Work-in-progress extract registration command to separate class
